Problem: #69 App architecture inefficiencies and overhead
Solution: Moved most of lib into imports. #69
