+++ imports/api/activityLog/activityLog.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const ActivityLog = new Mongo.Collection('activitylog');
