+++ imports/api/coins/changedCurrencies.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const ChangedCurrencies = new Mongo.Collection('changedcurrencies');
