+++ imports/api/coins/pendingCurrencies.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const PendingCurrencies = new Mongo.Collection('pendingcurrencies');
