+++ imports/api/coins/rejectedCurrencies.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const RejectedCurrencies = new Mongo.Collection('rejectedcurrencies');
