+++ imports/api/communities/communities.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const Communities = new Mongo.Collection('communities')
