+++ imports/api/developers/codebase.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const Codebase = new Mongo.Collection('codebase')
