+++ imports/api/developers/developers.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const Developers = new Mongo.Collection('developers')
