+++ imports/api/features/features.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const Features = new Mongo.Collection('features');
