+++ imports/api/miscellaneous/appLogs.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const AppLogs = new Mongo.Collection('AppLogs');
