+++ imports/api/ratings/eloRankings.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const EloRankings = new Mongo.Collection('elorankings');
