+++ imports/api/ratings/ratings.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Ratings = new Mongo.Collection('ratings');
