+++ imports/api/ratings/ratingsTemplates.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const RatingsTemplates = new Mongo.Collection('ratings_templates');
