+++ imports/api/redflags/methods.js
@@ -1,27 +1,5 @@
-export var Redflags = new Mongo.Collection('redflags');
-
-if(Meteor.isServer) {
-  Meteor.publish('redflags', function(id) {
-    if(!id) {
-      console.log("Redflags NoID");
-      return Redflags.find();
-    } else {
-      console.log("Redflags ID: " + id);
-      return Redflags.find({currencyId: id});
-    }
-  });
-
-  Meteor.publish('redflagcomments', function(id) {
-    if(!id) {
-      console.log("Redflags NoID");
-      return Redflags.find();
-    } else {
-      console.log("Redflags ID: " + id);
-      return Redflags.find({parentId: id});
-    }
-  });
-}
-
+import { Meteor } from 'meteor/meteor'
+import { Redflags } from '/imports/api/indexDB.js'
 
 Meteor.methods({
   redFlagVote: function(id, direction) {
