+++ imports/api/redflags/redflags.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const Redflags = new Mongo.Collection('redflags');
