+++ imports/api/users/profileImages.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo'
+
+export const ProfileImages = new Mongo.Collection('profileimages')
