+++ imports/api/wallet/wallet.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Wallet = new Mongo.Collection('wallet');
