+++ imports/api/wallet/walletImages.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const WalletImages = new Mongo.Collection('walletimages');
