+++ imports/startup/both/index.js
@@ -2,3 +2,7 @@ import '../../api/coins/methods.js'
 import '../../api/hashing/methods'
 import '../../api/users/methods'
 import '../../api/miscellaneous/methods'
+import '../../api/features/methods'
+import '../../api/redflags/methods'
+
+import './routes.js'
