+++ imports/startup/client/index.js
@@ -1,2 +1,1 @@
-import '../both/routes.js';
 import '../both/index.js'
