+++ imports/startup/server/index.js
@@ -5,8 +5,21 @@ import '/imports/api/bounties/server/publications'
 import '/imports/api/hashing/server/publications.js'
 import '/imports/api/miscellaneous/server/publications'
 import '/imports/api/users/server/publications'
+import '/imports/api/wallet/server/publications'
+import '/imports/api/developers/server/publications'
+import '/imports/api/ratings/server/publications'
+import '/imports/api/features/server/publications'
+import '/imports/api/communities/server/publications'
+import '/imports/api/coins/server/publications'
+import '/imports/api/redflags/server/publications'
+import '/imports/api/activityLog/server/publications'
+
+import '/imports/api/coins/server/methods'
+import '/imports/api/wallet/server/methods'
+import '/imports/api/activityLog/server/methods'
 
 // import './fastRender.js'
-import '../both/routes.js'
-import './users.js'
 import './fixtures.js'
+
+import '/imports/api/users/users.js'
+import '/imports/api/coins/currencies.js'
