--- lib/database/AppLogs.js
@@ -1,2 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-export var AppLogs = new Mongo.Collection('AppLogs');
