+++ lib/database/Bounties.js
@@ -3,6 +3,8 @@ if(Meteor.isServer) {
   import { log } from '../../server/main'
 }
 
+//THIS FILE WAS REPLICATED IN .../API/UTILITIES
+
 //inserted into methods
 export var REWARDCOEFFICIENT = 60000000;
