--- lib/database/Images.js
@@ -1,18 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-export var WalletImages = new Mongo.Collection('walletimages');
-
-if(Meteor.isServer){
-  Meteor.publish('walletimages', function walletimagesPublication(id) {
-
-  	if(id){
-  		return WalletImages.find({currencyId:id});
-  	}else{
-  		return WalletImages.find();
-  	}
-  
-  })
-
-  Meteor.publish('walletImagesSlug', slug => WalletImages.find({
-  	currencySlug: slug
-  }))
-}
